Add implementation of RFC 7959 (block-wise transfers)
This change was accompanied by a massive restructure and rewrite
of the messaging protocol flows.

* Implemented a comprehensive state machine for performing block-wise
  transfers in requests and responses with both block options in
  control and descriptive usages, block size negotiation, handling
  block-wise transfers of observe notifications and separate
  responses, handling many corner cases for fault tolerance
  and to accurately reflect the RFC.
* Requests and responses are processed atomically on arrival of all
  blocks from the exchange. Applications need have no knowledge of an
  ongoing block-wise transfer as the framework handles all of the logic
  internally.
* Buffers are internally allocated to hold state and are freed on
  completion of the exchange or after EXCHANGE_LIFETIME.
* The entire feature may be conditionally compiled in by setting
  OC_BLOCK_WISE_SET_MTU in the framework configuration. This asserts an
  MTU size limitation of the lower protocol layers, a maximum block size
  is appropriately computed and a block-wise transfer is automatically
  initiated if the request/response payload size exceeds this limit.
  Not setting OC_BLOCK_WISE_SET_MTU excludes the feature entirely
  (so lower code size) and resulting applications would be
  unable to handle block options (except they would reject requests
  containing them).
* Despite the existence of parallel flows (with/without block-wise
  transfer), there is no code duplication; the implementation is
  distributed over the set of appropriate modules with the parallel
  paths clearly demarcated by #ifdefs.
* Added client/server samples to exercise the block-wise transfer
  feature in both the request and response directions. Added the samples
  to .gitignore.
* Added a new framework configuration parameter OC_MAX_PDU_BUFFER_SIZE
  which sets the size of preallocated buffers to be able to
  accommodate the complete request/response payload.
* Defined a new macro OC_PDU_SIZE which will be the size of data buffers
  in the oc_message_t struct and will decide the maximum data size that may
  be sent to or received from the lower layers in the connectivity
  implementation.
* Updated oc_rep to fail prematurely on encountering CborErrorUnexpectedEOF
  during a parse so as to signal the remote endpoint that it has received
  incomplete data with a 4.13 response.
* Updated the Linux, Zephyr, Contiki and RIOT ports to adapt to the above
  changes.

+++ include/oc_rep.h
@@ -217,7 +217,7 @@ typedef struct oc_rep_s
   oc_string_t name;
   union
   {
-    int64_t value_int;
+    int value_int;
     bool value_boolean;
     double value_double;
     oc_string_t value_string;
